+++ demisto_sdk/__main__.py
@@ -1895,6 +1895,11 @@ def generate_test_playbook(ctx, **kwargs):
 @click.option(
     "--script", is_flag=True, help="Create a Script based on BaseScript example"
 )
+@click.option(
+    "--xsiam",
+    is_flag=True,
+    help="Create an Event Collector based on a template, and create matching sub directories",
+)
 @click.option("--pack", is_flag=True, help="Create pack and its sub directories")
 @click.option(
     "-t",
@@ -1929,7 +1934,8 @@ def init(ctx, **kwargs):
     from demisto_sdk.commands.init.initiator import Initiator
 
     check_configuration_file("init", kwargs)
-    initiator = Initiator(**kwargs)
+    marketplace = parse_marketplace_kwargs(kwargs)
+    initiator = Initiator(marketplace=marketplace, **kwargs)
     initiator.init()
     return 0
 
@@ -2089,12 +2095,14 @@ def _generate_docs_for_file(kwargs: Dict[str, Any]):
         if command:
             if (
                 output_path
-                and (not Path(output_path, "README.md").is_file())
+                and (not os.path.isfile(os.path.join(output_path, "README.md")))
                 or (not output_path)
                 and (
-                    not Path(
-                        os.path.dirname(os.path.realpath(input_path)), "README.md"
-                    ).is_file()
+                    not os.path.isfile(
+                        os.path.join(
+                            os.path.dirname(os.path.realpath(input_path)), "README.md"
+                        )
+                    )
                 )
             ):
                 raise Exception(
@@ -2112,7 +2120,7 @@ def _generate_docs_for_file(kwargs: Dict[str, Any]):
                 "[red]File is not an Integration, Script, Playbook or a README.[/red]"
             )
 
-        if old_version and not Path(old_version).is_file():
+        if old_version and not os.path.isfile(old_version):
             raise Exception(
                 f"[red]Input old version file {old_version} was not found.[/red]"
             )
@@ -2679,7 +2687,7 @@ def openapi_codegen(ctx, **kwargs):
         output_dir = kwargs["output_dir"]
 
     # Check the directory exists and if not, try to create it
-    if not Path(output_dir).exists():
+    if not os.path.exists(output_dir):
         try:
             os.mkdir(output_dir)
         except Exception as err:
@@ -2902,8 +2910,8 @@ def test_content(ctx, **kwargs):
     default=False,
 )
 @click.option(
-    "-pkw",
-    "--use-packs-known-words",
+    "-pkw/-spkw",
+    "--use-packs-known-words/--skip-packs-known-words",
     is_flag=True,
     help="Will find and load the known_words file from the pack. "
     "To use this option make sure you are running from the "