+++ demisto_sdk/commands/lint/lint_manager.py
@@ -293,9 +293,8 @@ class LintManager:
             if isinstance(input, str):
                 input = input.split(",")
             for item in input:
-                is_pack = (
-                    os.path.isdir(item)
-                    and Path(item, PACKS_PACK_META_FILE_NAME).exists()
+                is_pack = os.path.isdir(item) and os.path.exists(
+                    os.path.join(item, PACKS_PACK_META_FILE_NAME)
                 )
                 if is_pack:
                     pkgs.extend(LintManager._get_all_packages(content_dir=item))
@@ -1176,7 +1175,7 @@ class LintManager:
         if not self.json_file_path:
             return
 
-        if Path(self.json_file_path).exists():
+        if os.path.exists(self.json_file_path):
             json_contents = get_json(self.json_file_path)
             if not (isinstance(json_contents, list)):
                 json_contents = []
@@ -1237,7 +1236,7 @@ class LintManager:
         mypy_errors: list = []
         gather_error: list = []
         for line in error_messages:
-            if Path(line.split(":")[0]).is_file():
+            if os.path.isfile(line.split(":")[0]):
                 if gather_error:
                     mypy_errors.append("\n".join(gather_error))
                     gather_error = []