+++ demisto_sdk/commands/prepare_content/rule_unifier.py
@@ -56,7 +56,7 @@ class RuleUnifier(Unifier):
     @staticmethod
     def _insert_schema(path: Path, data: dict):
         schema_path = str(path).replace(".yml", "_schema.json")
-        if Path(schema_path).exists():
+        if os.path.exists(schema_path):
             with open(schema_path, encoding="utf-8") as schema_file:
                 schema = json.loads(schema_file.read())
                 data["schema"] = FoldedScalarString(json.dumps(schema, indent=4))