+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Changelog
 ## Unreleased
+* Added the *--xsiam* flag to the **init** command in order to create XSIAM content.
+
+## 1.18.1
 * Fixed an issue where the coloring directives where showing in log messages.
 * Fixed an issue where **create-content-graph** was not executed upon changes in the parser infra files.
+<<<<<<< HEAD
 * Updated the **format** command to use the content graph instead of the id_set file. 
+=======
+* Added a parameter `skip-packs-known-words` to the **doc-review** command, making sure that pack known words will not be added.
+>>>>>>> master
 
 ## 1.18.0
 * Added the ability to ignore any validation in the **validate** command when running in an external (non-demisto/content) repo, by placing a `.private-repo-settings` file at its root.